requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.38
psycopg[binary]==3.2.13
alembic==1.13.1

# Config & Environment
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Authentication & Security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0

# HTTP / OAuth
httpx==0.27.0
authlib==1.3.0

# Utilities
email-validator==2.1.0.post1
